requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
yfinance==0.2.40
requests==2.32.3
SQLAlchemy==2.0.36
openpyxl==3.1.5
matplotlib==3.9.2
pyarrow==17.0.0
